+++ booking-service/internal/booking/event_client.go
@@ -3,5 +3,5 @@ package booking
 import "context"
 
 type EventClientService interface {
-	ReserveSeats(ctx context.Context, eventPublicID string, seats uint64) (float64, error)
+	ReserveSeats(ctx context.Context, eventPublicID string, seats uint64) (uint64, error)
 }
